Lex kent SQL niet + bug fix

+++ src/main/java/nl/avans/wordcrex/model/User.java
@@ -348,6 +348,10 @@ public class User implements Pollable<User> {
         }
 
         if (user.hasRole(role)) {
+            if (user.roles.size() <= 1) {
+                return;
+            }
+
             this.database.update(
                 "DELETE FROM accountrole WHERE role = ? AND username = ?",
                 (statement) -> {
